requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic[email]==2.5.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.6
pydantic-settings==2.1.0
